requirements.txt
Flask
flask-restx
requests
python-dotenv